refactor: Split tests/it/others.rs
